+++ 03/03-03.py
@@ -1,7 +1,10 @@
+#Input för Ord.Pris vilket sedan intergeras för att kunna användas
 ordPris = input("Ange ordinarie pris: ")
-ordPris = int(ordPris)
+ordPris = float(ordPris)
 
+#uträkning av extrapriset
 extrapris = ordPris * 0.85
 int(extrapris)
 
+#utskrift på vad som ska betalas 
 print("Pris med rabatt: {0} kr".format(extrapris))
